+++ browser-switch/src/test/java/com/braintreepayments/api/BrowserSwitchClientUnitTest.java
@@ -1,18 +1,14 @@
 package com.braintreepayments.api;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertSame;
 import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.ArgumentMatchers.same;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -23,15 +19,11 @@ import android.content.Intent;
 import android.net.Uri;
 
 import androidx.activity.ComponentActivity;
-import androidx.fragment.app.FragmentActivity;
 
 import org.json.JSONObject;
 import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
-import org.mockito.ArgumentCaptor;
 import org.robolectric.Robolectric;
 import org.robolectric.RobolectricTestRunner;
 import org.robolectric.android.controller.ActivityController;
@@ -39,10 +31,6 @@ import org.robolectric.android.controller.ActivityController;
 @RunWith(RobolectricTestRunner.class)
 public class BrowserSwitchClientUnitTest {
 
-    @Rule
-    public ExpectedException exceptionRule = ExpectedException.none();
-
-    private BrowserSwitchPersistentStore persistentStore;
     private BrowserSwitchInspector browserSwitchInspector;
 
     private ChromeCustomTabsInternalClient customTabsInternalClient;
@@ -50,27 +38,20 @@ public class BrowserSwitchClientUnitTest {
     private Uri browserSwitchDestinationUrl;
     private Context applicationContext;
 
-    private ActivityController<FragmentActivity> activityController;
-    private FragmentActivity activity;
-
-    private ActivityController<ComponentActivity> componentActivityController;
     private ComponentActivity componentActivity;
+
     @Before
     public void beforeEach() {
-        persistentStore = mock(BrowserSwitchPersistentStore.class);
-
         browserSwitchInspector = mock(BrowserSwitchInspector.class);
         customTabsInternalClient = mock(ChromeCustomTabsInternalClient.class);
 
         browserSwitchDestinationUrl = Uri.parse("https://example.com/browser_switch_destination");
 
-        activityController = Robolectric.buildActivity(FragmentActivity.class).setup();
-        activity = spy(activityController.get());
-
-        componentActivityController = Robolectric.buildActivity(ComponentActivity.class).setup();
+        ActivityController<ComponentActivity> componentActivityController =
+                Robolectric.buildActivity(ComponentActivity.class).setup();
         componentActivity = spy(componentActivityController.get());
 
-        applicationContext = activity.getApplicationContext();
+        applicationContext = componentActivity.getApplicationContext();
     }
 
     @Test
@@ -79,7 +60,8 @@ public class BrowserSwitchClientUnitTest {
 
         when(componentActivity.isFinishing()).thenReturn(true);
 
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
+        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector,
+                customTabsInternalClient);
 
         JSONObject metadata = new JSONObject();
         BrowserSwitchOptions options = new BrowserSwitchOptions()
@@ -97,7 +79,8 @@ public class BrowserSwitchClientUnitTest {
     public void start_whenSuccessful_returnsBrowserSwitchRequest() {
         when(browserSwitchInspector.isDeviceConfiguredForDeepLinking(applicationContext, "return-url-scheme")).thenReturn(true);
 
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
+        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector,
+                customTabsInternalClient);
 
         JSONObject metadata = new JSONObject();
         BrowserSwitchOptions options = new BrowserSwitchOptions()
@@ -124,7 +107,8 @@ public class BrowserSwitchClientUnitTest {
         when(browserSwitchInspector.isDeviceConfiguredForDeepLinking(applicationContext, "return-url-scheme")).thenReturn(true);
         doThrow(new ActivityNotFoundException()).when(customTabsInternalClient).launchUrl(any(Context.class), any(Uri.class), eq(false));
 
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
+        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector,
+                customTabsInternalClient);
 
         JSONObject metadata = new JSONObject();
         BrowserSwitchOptions options = new BrowserSwitchOptions()
@@ -140,7 +124,8 @@ public class BrowserSwitchClientUnitTest {
     public void start_whenRequestCodeIsIntegerMinValue_returnsFailure() {
         when(browserSwitchInspector.isDeviceConfiguredForDeepLinking(applicationContext, "return-url-scheme")).thenReturn(true);
 
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
+        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector,
+                customTabsInternalClient);
 
         JSONObject metadata = new JSONObject();
         BrowserSwitchOptions options = new BrowserSwitchOptions()
@@ -157,7 +142,8 @@ public class BrowserSwitchClientUnitTest {
     public void start_whenDeviceIsNotConfiguredForDeepLinking_returnsFailure() {
         when(browserSwitchInspector.isDeviceConfiguredForDeepLinking(applicationContext, "return-url-scheme")).thenReturn(false);
 
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
+        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector,
+                customTabsInternalClient);
 
         JSONObject metadata = new JSONObject();
         BrowserSwitchOptions options = new BrowserSwitchOptions()
@@ -178,7 +164,8 @@ public class BrowserSwitchClientUnitTest {
     public void start_whenNoReturnUrlSchemeSet_throwsFailure() {
         when(browserSwitchInspector.isDeviceConfiguredForDeepLinking(applicationContext, "return-url-scheme")).thenReturn(true);
 
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
+        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector,
+                customTabsInternalClient);
 
         JSONObject metadata = new JSONObject();
         BrowserSwitchOptions options = new BrowserSwitchOptions()
@@ -192,200 +179,9 @@ public class BrowserSwitchClientUnitTest {
     }
 
     @Test
-    public void deliverResult_whenDeepLinkUrlExistsAndReturnUrlSchemeMatchesAndNoPendingRequest_returnsNull() {
-        when(activity.getApplicationContext()).thenReturn(applicationContext);
-
-        Uri deepLinkUrl = Uri.parse("return-url-scheme://test");
-        Intent deepLinkIntent = new Intent().setData(deepLinkUrl);
-        when(activity.getIntent()).thenReturn(deepLinkIntent);
-
-        when(persistentStore.getActiveRequest(applicationContext)).thenReturn(null);
-
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
-        BrowserSwitchResult result = sut.deliverResult(activity);
-
-        assertNull(result);
-    }
-
-    @Test
-    public void deliverResult_whenDeepLinkUrlExists_AndReturnUrlSchemeDoesNotMatch_AndShouldNotifyCancellation_notifiesResultCANCELED_AndSetsRequestShouldNotifyCancellationToFalse() {
-        when(activity.getApplicationContext()).thenReturn(applicationContext);
-
-        Uri deepLinkUrl = Uri.parse("another-return-url-scheme://test");
-        Intent deepLinkIntent = new Intent().setData(deepLinkUrl);
-        when(activity.getIntent()).thenReturn(deepLinkIntent);
-
-        JSONObject requestMetadata = new JSONObject();
-        BrowserSwitchRequest request = new BrowserSwitchRequest(123, browserSwitchDestinationUrl, requestMetadata, "return-url-scheme", true);
-        when(persistentStore.getActiveRequest(applicationContext)).thenReturn(request);
-
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
-        BrowserSwitchResult result = sut.deliverResult(activity);
-
-        assertNotNull(result);
-        assertEquals(123, result.getRequestCode());
-        assertSame(browserSwitchDestinationUrl, result.getRequestUrl());
-        assertEquals(result.getStatus(), BrowserSwitchStatus.CANCELED);
-        assertSame(result.getRequestMetadata(), requestMetadata);
-        assertNull(result.getDeepLinkUrl());
-
-        ArgumentCaptor<BrowserSwitchRequest> captor = ArgumentCaptor.forClass(BrowserSwitchRequest.class);
-        verify(persistentStore).putActiveRequest(captor.capture(), same(activity));
-
-        BrowserSwitchRequest updatedRequest = captor.getValue();
-        assertSame(request, updatedRequest);
-        assertFalse(updatedRequest.getShouldNotifyCancellation());
-    }
-
-    @Test
-    public void deliverResult_whenDeepLinkUrlExistsAndReturnUrlSchemeDoesNotMatchAndShouldNotNotifyCancellation_returnsNull() {
-        when(activity.getApplicationContext()).thenReturn(applicationContext);
-
-        Uri deepLinkUrl = Uri.parse("another-return-url-scheme://test");
-        Intent deepLinkIntent = new Intent().setData(deepLinkUrl);
-        when(activity.getIntent()).thenReturn(deepLinkIntent);
-
-        JSONObject requestMetadata = new JSONObject();
-        BrowserSwitchRequest request = new BrowserSwitchRequest(123, browserSwitchDestinationUrl, requestMetadata, "return-url-scheme", false);
-        when(persistentStore.getActiveRequest(applicationContext)).thenReturn(request);
-
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
-        BrowserSwitchResult result = sut.deliverResult(activity);
-
-        assertNull(result);
-        verify(persistentStore, never()).putActiveRequest(any(BrowserSwitchRequest.class), any(FragmentActivity.class));
-    }
-
-    @Test
-    public void deliverResult_whenDeepLinkUrlDoesNotExistAndShouldNotifyCancellation_notifiesResultCANCELEDAndSetsRequestShouldNotifyCancellationToFalse() {
-        when(activity.getApplicationContext()).thenReturn(applicationContext);
-        when(activity.getIntent()).thenReturn(new Intent());
-
-        JSONObject requestMetadata = new JSONObject();
-        BrowserSwitchRequest request =
-                new BrowserSwitchRequest(123, browserSwitchDestinationUrl, requestMetadata, "return-url-scheme", true);
-        when(persistentStore.getActiveRequest(applicationContext)).thenReturn(request);
-
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
-        BrowserSwitchResult result = sut.deliverResult(activity);
-
-        assertNotNull(result);
-        assertEquals(123, result.getRequestCode());
-        assertSame(browserSwitchDestinationUrl, result.getRequestUrl());
-        assertEquals(result.getStatus(), BrowserSwitchStatus.CANCELED);
-        assertSame(result.getRequestMetadata(), requestMetadata);
-        assertNull(result.getDeepLinkUrl());
-
-        ArgumentCaptor<BrowserSwitchRequest> captor = ArgumentCaptor.forClass(BrowserSwitchRequest.class);
-        verify(persistentStore).putActiveRequest(captor.capture(), same(activity));
-
-        BrowserSwitchRequest updatedRequest = captor.getValue();
-        assertSame(request, updatedRequest);
-        assertFalse(updatedRequest.getShouldNotifyCancellation());
-    }
-
-    @Test
-    public void deliverResult_whenDeepLinkUrlDoesNotExistAndShouldNotNotifyCancellation_returnsNull() {
-        when(activity.getApplicationContext()).thenReturn(applicationContext);
-        when(activity.getIntent()).thenReturn(new Intent());
-
-        JSONObject requestMetadata = new JSONObject();
-        BrowserSwitchRequest request =
-                new BrowserSwitchRequest(123, browserSwitchDestinationUrl, requestMetadata, "return-url-scheme", false);
-        when(persistentStore.getActiveRequest(applicationContext)).thenReturn(request);
-
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
-        BrowserSwitchResult result = sut.deliverResult(activity);
-
-        assertNull(result);
-        verify(persistentStore, never()).putActiveRequest(any(BrowserSwitchRequest.class), any(FragmentActivity.class));
-    }
-
-    @Test
-    public void deliverResult_whenRequestIsNull_doesNothing() {
-        when(activity.getApplicationContext()).thenReturn(applicationContext);
-        when(persistentStore.getActiveRequest(applicationContext)).thenReturn(null);
-
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
-        BrowserSwitchResult result = sut.deliverResult(activity);
-
-        assertNull(result);
-        verify(persistentStore, never()).clearActiveRequest(activity);
-    }
-
-    @Test
-    public void captureResult_whenDeepLinkUrlExistsAndReturnUrlSchemeMatches_storesSuccessResultInPersistentStore() {
-        when(activity.getApplicationContext()).thenReturn(applicationContext);
-
-        Uri deepLinkUrl = Uri.parse("return-url-scheme://test");
-        Intent deepLinkIntent = new Intent().setData(deepLinkUrl);
-        when(activity.getIntent()).thenReturn(deepLinkIntent);
-
-        JSONObject requestMetadata = new JSONObject();
-        BrowserSwitchRequest request = new BrowserSwitchRequest(123, browserSwitchDestinationUrl, requestMetadata, "return-url-scheme", false);
-        when(persistentStore.getActiveRequest(applicationContext)).thenReturn(request);
-
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
-        sut.captureResult(activity);
-
-        ArgumentCaptor<BrowserSwitchResult> captor =
-                ArgumentCaptor.forClass(BrowserSwitchResult.class);
-        verify(persistentStore).putActiveResult(captor.capture(), same(applicationContext));
-
-        BrowserSwitchResult result = captor.getValue();
-        assertNotNull(result);
-        assertEquals(123, result.getRequestCode());
-        assertSame(browserSwitchDestinationUrl, result.getRequestUrl());
-        assertEquals(BrowserSwitchStatus.SUCCESS, result.getStatus());
-        assertSame(requestMetadata, result.getRequestMetadata());
-        assertSame(deepLinkUrl, result.getDeepLinkUrl());
-    }
-
-    @Test
-    public void captureResult_whenRequestIsNull_doesNothing() {
-        when(activity.getApplicationContext()).thenReturn(applicationContext);
-        when(persistentStore.getActiveRequest(applicationContext)).thenReturn(null);
-
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
-        sut.captureResult(activity);
-
-        verify(persistentStore, never()).putActiveResult(any(BrowserSwitchResult.class), any(Context.class));
-    }
-
-    @Test
-    public void captureResult_whenIntentIsNull_doesNothing() {
-        when(activity.getApplicationContext()).thenReturn(applicationContext);
-        when(activity.getIntent()).thenReturn(null);
-
-        JSONObject requestMetadata = new JSONObject();
-        BrowserSwitchRequest request = new BrowserSwitchRequest(123, browserSwitchDestinationUrl, requestMetadata, "return-url-scheme", false);
-        when(persistentStore.getActiveRequest(applicationContext)).thenReturn(request);
-
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
-        sut.captureResult(activity);
-
-        verify(persistentStore, never()).putActiveResult(any(BrowserSwitchResult.class), any(Context.class));
-    }
-
-    @Test
-    public void deliverResultFromCache_forwardsCachedResultFromBrowserSwitchPersistentStorageAndRemovesAllItemsFromPersistentStorage() {
-        when(activity.getApplicationContext()).thenReturn(applicationContext);
-
-        JSONObject requestMetadata = new JSONObject();
-        BrowserSwitchRequest request = new BrowserSwitchRequest(123, browserSwitchDestinationUrl, requestMetadata, "return-url-scheme", false);
-        BrowserSwitchResult cachedResult = new BrowserSwitchResult(BrowserSwitchStatus.SUCCESS, request, Uri.parse("example://success/url"));
-        when(persistentStore.getActiveResult(same(applicationContext))).thenReturn(cachedResult);
-
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
-        BrowserSwitchResult actualResult = sut.deliverResultFromCache(activity);
-
-        assertSame(cachedResult, actualResult);
-        verify(persistentStore).removeAll(applicationContext);
-    }
-
-    @Test
     public void parseResult_whenActiveRequestMatchesDeepLinkResultURLScheme_returnsBrowserSwitchSuccessResult() {
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
+        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector,
+                customTabsInternalClient);
 
         JSONObject requestMetadata = new JSONObject();
         BrowserSwitchRequest request =
@@ -402,7 +198,8 @@ public class BrowserSwitchClientUnitTest {
 
     @Test
     public void parseResult_whenDeepLinkResultURLSchemeDoesntMatch_returnsNull() {
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
+        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector,
+                customTabsInternalClient);
 
         JSONObject requestMetadata = new JSONObject();
         BrowserSwitchRequest request =
@@ -417,7 +214,8 @@ public class BrowserSwitchClientUnitTest {
 
     @Test
     public void parseResult_whenIntentIsNull_returnsNull() {
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
+        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector,
+                customTabsInternalClient);
 
         JSONObject requestMetadata = new JSONObject();
         BrowserSwitchRequest request =
@@ -425,13 +223,5 @@ public class BrowserSwitchClientUnitTest {
 
         BrowserSwitchResult browserSwitchResult = sut.parseResult(new BrowserSwitchPendingRequest.Started(request), null);
         assertNull(browserSwitchResult);
-    }
-
-    @Test
-    public void clearActiveRequests_forwardsInvocationToPersistantStore() {
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
-
-        sut.clearActiveRequests(applicationContext);
-        verify(persistentStore).clearActiveRequest(applicationContext);
     }
 }